+++ src/app/app.module.ts
@@ -6,7 +6,7 @@ import { NgModule } from '@angular/core';
 import {
     IgxIconModule, IgxGridModule, IgxExcelExporterService, IgxCsvExporterService, IgxOverlayService,
     IgxDragDropModule, IgxDividerModule, IgxTreeGridModule,  IgxHierarchicalGridModule, IgxInputGroupModule,
-    IgxIconService, DisplayDensityToken, DisplayDensity, IgxDateTimeEditorModule, IgxButtonModule, IgxDateRangeModule
+    IgxIconService, DisplayDensityToken, DisplayDensity, IgxDateTimeEditorModule, IgxButtonModule, IgxDateRangePickerModule
 } from 'igniteui-angular';
 import { IgxColumnHidingModule } from 'igniteui-angular';
 import { SharedModule } from './shared/shared.module';
@@ -254,7 +254,7 @@ const components = [
         IgxHierarchicalGridModule,
         IgxColumnHidingModule,
         IgxDragDropModule,
-        IgxDateRangeModule,
+        IgxDateRangePickerModule,
         IgxDividerModule,
         SharedModule,
         routing,